refactor(project): optimize code structure of the flat-web (#1613)
* refactor(project): optimize code structure of the flat-web

* chore(flat-pages): move file
