+++ packages/flat-pages/src/assets/image/default-avatars/0.png
