+++ packages/flat-pages/src/assets/image/default-avatars/1.png
