+++ packages/flat-pages/src/assets/image/default-avatars/2.png
