+++ packages/flat-pages/src/assets/image/default-avatars/3.png
