+++ packages/flat-pages/src/assets/image/default-avatars/4.png
